Updated for XUbuntu 22.04

+++ base/exfat_support.sh
@@ -12,7 +12,14 @@ if [[ "$1" == "--help" || "$1" == "-h" ]]; then
 fi
 
 #==============================================================================
-_title "Installing ExFAT support on your computer"
+_title "Adding universe to apt repository..."
 #==============================================================================
 add-apt-repository -y universe && apt update
-apt install -y exfat-fuse exfat-utils
+
+#==============================================================================
+# If we are running Ubuntu 22.04 variants, we don't need this:
+	#==============================================================================
+if [[ "$(cat /etc/os-release | grep "VERSION=" | cut -d"\"" -f 2 | cut -d" " -f 1)" != "22.04" ]]; then
+	_title "Installing ExFAT support on your computer"
+	apt install -y exfat-fuse exfat-utils
+fi
